+++ src/main/java/com/starkbank/Transfer.java
@@ -163,6 +163,7 @@ public final class Transfer extends Resource {
      * before [string, default null]: date filter for objects created or updated only before specified date. ex: "2020-03-10"
      * transactionIds [list of strings, default null]: list of transaction IDs linked to the desired transfers. ex: ["5656565656565656", "4545454545454545"]
      * status [string, default null]: filter for status of retrieved objects. ex: "processing" or "success"
+     * taxId [string, default null]: filter for transfers sent to the specified tax ID. ex: "012.345.678-90"
      * sort [string, default "-created"]: sort order considered in response. Valid options are "created", "-created", "updated" or "-updated".
      * tags [list of strings, default null]: tags to filter retrieved objects. ex: ["tony", "stark"]
      * <p>