+++ src/main/java/com/starkbank/error/ErrorElement.java
@@ -1,7 +1,7 @@
 package com.starkbank.error;
 
 public final class ErrorElement extends Error{
-	private static final long serialVersionUID = 5417470047486626244L;
+
 	public String code;
     public String message;