+++ src/main/java/com/starkbank/error/InternalServerError.java
@@ -1,7 +1,6 @@
 package com.starkbank.error;
 
 public final class InternalServerError extends RuntimeException{
-    private static final long serialVersionUID = 4458611990624094212L;
 
     public InternalServerError(String message){
         super(message);