+++ src/main/java/com/starkbank/error/UnknownError.java
@@ -1,0 +1,8 @@
+package com.starkbank.error;
+
+public final class UnknownError extends RuntimeException{
+
+    public UnknownError(String message){
+        super(message);
+    }
+}